+++ diesel/src/mysql/query_builder/query_fragment_impls.rs
@@ -1,0 +1,11 @@
+use mysql::Mysql;
+use query_builder::{AstPass, QueryFragment};
+use query_builder::for_update_clause::ForUpdateClause;
+use result::QueryResult;
+
+impl QueryFragment<Mysql> for ForUpdateClause {
+    fn walk_ast(&self, mut out: AstPass<Mysql>) -> QueryResult<()> {
+        out.push_sql(" FOR UPDATE");
+        Ok(())
+    }
+}